+++ src/components/Changelog.tsx
@@ -19,6 +19,20 @@ const Changelog: React.FC = () => {
             <li>Brand new layout system</li>
           </ul>
         </li>
+        <li>
+          <h4>0.2.0 (2019-12-7)</h4>
+          <ul>
+            <li>
+              Fix painting inconsistencies between the preview and the current
+              layer
+            </li>
+            <li>
+              Improve Welcome window, making it easier to close, just by
+              clicking outside of it
+            </li>
+            <li>Add a small padding to the default size of the canvas</li>
+          </ul>
+        </li>
       </ul>
 
       <p style={{ textAlign: 'center' }}>